Rename evaluate_preds module to evaluation. This is a clearer name for this module.
